refactor: move BlockKeywordMatcher to utils.bsl package
BSL-специфичная утилита по поиску парных ключевых слов, не имеет
зависимостей от LSP/provider-инфраструктуры. Логичное место —
utils.bsl (рядом с Constructors), а не providers.
